fix: resolve all MSVC compiler warnings (C4244, C4245, C4100, C4305, C4505)
- C4305: Add 'f' suffix to double literals in futils.h constexpr float constants
- C4245: Use ~uint32_t(0) instead of -1, explicit uint32_t casts for SIMD init
- C4100: Comment out unreferenced parameter names in virtual overrides
- C4244: Add explicit static_cast<float>/int/mono_float for narrowing conversions
- C4505: Mark toggleFilter1/toggleFilter2 as [[maybe_unused]]
- CMakeLists.txt: Switch Windows installer from Inno Setup to WiX (matches project CI)

+++ src/dsp/Chorus.cpp
@@ -63,7 +63,7 @@ int Chorus::getNextNumVoicePairs() {
 // Mirrors vial::ChorusModule::processWithInput (chorus_module.cpp:92-156).
 void Chorus::process (const poly_float* audio_in, int num_samples) {
     poly_float frequency = frequencyHz_;
-    poly_float delta_phase = (frequency * num_samples) * (1.0f / sampleRate_);
+    poly_float delta_phase = (frequency * static_cast<float>(num_samples)) * (1.0f / static_cast<float>(sampleRate_));
     phase_ = utils::mod (phase_ + delta_phase);
 
     poly_float* audio_out = output_.buffer;
